W160k: y_runtime 0044AD20 568/576 nd349 -> 572/576 MATCH (YVec3f aggregate copy + absolute allocator/data aliases + 32-bit result lifetimes + guard-return flow + signed byte views + persistent-local order); 00457470 452/368 nd303 -> 356/368 nd6, window CLEARED (absolute 0x1c0 row base, ordered mixed float-tail constant aggregate, call-result width, typed rotate call)

+++ src/Scene/mt_sceneFunc.c
@@ -37,7 +37,7 @@ extern u32 DAT_007cad5c;
 extern u32 DAT_007cada0;
 extern u32 DAT_007caea4;
 extern u32 DAT_007caf1c;
-extern u32 DAT_007cb0b8;
+extern f32 DAT_007cb0b8;
 extern u32 DAT_007cd560;
 extern u32 DAT_007cd564;
 extern u32 DAT_007ce658;
